Added a ModelDetailView to perform common function of detailviews
