+++ core/views/GearViews.py
@@ -4,24 +4,12 @@ from django.utils import timezone
 
 from core.models.GearModels import Gear
 from core.views.ViewList import RestrictedViewList
+from core.views.common import ModelDetailView
 
-class GearDetailView(DetailView):
+
+class GearDetailView(ModelDetailView):
 
     model = Gear
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context['now'] = timezone.now()
-        context['opts'] = self.model._meta
-        context['app_label'] = self.model._meta.app_label
-        context['change'] = False
-        context['is_popup'] = False
-        context['add'] = False
-        context['save_as'] = False
-        context['has_delete_permission'] = False
-        context['has_change_permission'] = False
-        context['has_add_permission'] = False
-        return context
 
 
 class GearViewList(RestrictedViewList):
