--- core/views/__init__.py
@@ -1,16 +1,0 @@
-from django.utils import timezone
-
-
-def get_default_context(obj, context):
-    """Convenience function for getting the default context data of a member"""
-    context['now'] = timezone.now()
-    context['opts'] = obj.model._meta
-    context['app_label'] = obj.model._meta.app_label
-    context['change'] = False
-    context['is_popup'] = False
-    context['add'] = False
-    context['save_as'] = False
-    context['has_delete_permission'] = False
-    context['has_change_permission'] = False
-    context['has_add_permission'] = False
-    return context
